refactor: Move AppVersion data class to domain module
Move the pure Kotlin AppVersion data class from androidApp/ to
domain/src/commonMain/. The Context.AppVersion() extension stays in
androidApp (reads PackageManager + BuildConfig — platform-specific).

Completes a piece of Phase 37 (KMP expect/actual) — Version info types
are now in commonMain, platform-specific version lookup stays in
androidApp.
